+++ CHANGELOG.md
@@ -12,6 +12,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Fixed `libcasm.clusterography.equivalents_info_from_dict`, which was trying to read clusters from the wrong position. This method was not used to read equivalents_info.json for KMC simulations.
 - Fixed `libcasm.occ_events.get_occevent_coordinate`, which was using the reverse of the appropriate translation to get the unitcell_index.
+- Fixed `to_json(to_json(config::Configuration)` and `read_dof_values` to properly support writing and reading DoF values in both the prim and standard basis.
 
 ### Changed